Initial - PE for Dev/eBook/requirements.txt
# Requirements for HTML to PDF conversion with fixed page numbering
weasyprint>=61.0
PyPDF2>=3.0.0
reportlab>=4.0.0

# Requirements for HTML to DOCX conversion
python-docx>=0.8.11
beautifulsoup4>=4.9.3
cssutils>=2.3.0
lxml>=4.6.0

# WeasyPrint dependencies are automatically handled
# but you may need system dependencies on some platforms:
# 
# On macOS with Homebrew:
# brew install python3 cairo pango gdk-pixbuf libffi
#
# On Ubuntu/Debian:
# sudo apt-get install python3-dev python3-pip python3-cffi python3-brotli libpango-1.0-0 libharfbuzz0b libpangoft2-1.0-0
#
# On CentOS/RHEL/Fedora:
# sudo yum install python3-devel python3-pip python3-cffi python3-brotli pango libffi-devel
